Fixed bug #852
 Jesse Anders      2009-10-17 14:17:25 PDT

In VC++, the build script that creates SDL_config.h and SDL_revision.h fails
when the project directory path includes spaces. For example, if the project
directory path is:

c:\Documents and Settings\Jesse\My Documents\Visual Studio 2008\Projects\SDL
1.3\VisualC\SDL\

The following error is generated:

'and' is not recognized as an internal or external command

Placing all paths in quotes fixes this.


git-svn-id: http://svn.libsdl.org/trunk/SDL@5125 c70aab31-4412-0410-b14c-859654838e24